prototyping showing coinjoin status all the time.
