+++ WalletWasabi.Gui/Controls/WalletExplorer/CoinJoinStatusView.xaml.cs
@@ -1,0 +1,18 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Gui.Controls.WalletExplorer
+{
+	public class CoinJoinStatusView : UserControl
+	{
+		public CoinJoinStatusView()
+		{
+			InitializeComponent();
+		}
+
+		private void InitializeComponent()
+		{
+			AvaloniaXamlLoader.Load(this);
+		}
+	}
+}
