+++ WalletWasabi.Gui/Controls/WalletExplorer/CoinJoinTabView.xaml
@@ -34,61 +34,7 @@
         <ColumnDefinition Width="150" />
         <ColumnDefinition Width="*" MinWidth="460" />
       </Grid.ColumnDefinitions>
-      <controls:GroupBox Grid.Row="0" Grid.Column="2" Title="Status" Padding="5" Margin="5 5 0 0">
-        <ScrollViewer>
-          <StackPanel Orientation="Vertical">
-            <StackPanel Orientation="Horizontal" Spacing="4">
-              <TextBlock Text="Minimum" />
-              <TextBlock Foreground="YellowGreen" Text="{Binding RequiredBTC, Converter={StaticResource MoneyStringConverter}}" />
-              <TextBlock Foreground="YellowGreen" Text="BTC" />
-              <TextBlock Text="is required to be queued for CoinJoin." />
-            </StackPanel>
-            <StackPanel Orientation="Horizontal" Spacing="4" IsVisible="{Binding AmountQueued, Converter={StaticResource CoinJoinedVisibilityConverter}}">
-              <TextBlock Text="You have queued" />
-              <TextBlock Foreground="YellowGreen" IsVisible="{Binding IsLurkingWifeMode, Converter={StaticResource InverseBooleanConverter}}" Text="{Binding AmountQueued, Converter={StaticResource MoneyStringConverter}}" />
-              <TextBlock Foreground="YellowGreen" IsVisible="{Binding IsLurkingWifeMode}" Text="{Binding AmountQueued, ConverterParameter=8, Converter={StaticResource LurkingWifeModeStringConverter}, Mode=OneWay}" />
-              <TextBlock Foreground="YellowGreen" Text="BTC" />
-              <TextBlock Text="for CoinJoin." />
-            </StackPanel>
-            <StackPanel Orientation="Horizontal" Spacing="4">
-              <TextBlock Text="Coordinator Fee:" />
-              <TextBlock Text="{Binding CoordinatorFeePercent}" />
-              <TextBlock Text="% per anonymity set." />
-            </StackPanel>
-            <Grid ColumnDefinitions="210, *" RowDefinitions="50, 20, 20">
-              <StackPanel Grid.Row="0" Grid.Column="0" Orientation="Horizontal" Spacing="4" Margin="0 12">
-                <Canvas Height="20" Width="200">
-                  <Ellipse Width="20" Height="20" StrokeThickness="4" Stroke="{Binding Phase, ConverterParameter=InputRegistration, Converter={StaticResource PhaseColorConverter}}" Canvas.Left="0" />
-                  <Ellipse Width="20" Height="20" StrokeThickness="4" Stroke="{Binding Phase, ConverterParameter=ConnectionConfirmation, Converter={StaticResource PhaseColorConverter}}" Canvas.Left="50" />
-                  <Ellipse Width="20" Height="20" StrokeThickness="4" Stroke="{Binding Phase, ConverterParameter=OutputRegistration, Converter={StaticResource PhaseColorConverter}}" Canvas.Left="100" />
-                  <Ellipse Width="20" Height="20" StrokeThickness="4" Stroke="{Binding Phase, ConverterParameter=Signing, Converter={StaticResource PhaseColorConverter}}" Canvas.Left="150" />
-                  <Line StartPoint="20,10" EndPoint="50,10" StrokeThickness="4" Stroke="{Binding Phase, ConverterParameter=InputRegistration, Converter={StaticResource PhaseColorConverter}}" />
-                  <Line StartPoint="70,10" EndPoint="100,10" StrokeThickness="4" Stroke="{Binding Phase, ConverterParameter=ConnectionConfirmation, Converter={StaticResource PhaseColorConverter}}" />
-                  <Line StartPoint="120,10" EndPoint="150,10" StrokeThickness="4" Stroke="{Binding Phase, ConverterParameter=OutputRegistration, Converter={StaticResource PhaseColorConverter}}" />
-                </Canvas>
-              </StackPanel>
-              <StackPanel Grid.Row="0" Grid.Column="1" Orientation="Horizontal" Spacing="4" Margin="0 12">
-                <TextBlock VerticalAlignment="Center" Foreground="YellowGreen" Text="{Binding Phase, Converter={StaticResource PhaseStringConverter}}" />
-              </StackPanel>
-              <StackPanel Grid.Row="1" Grid.Column="0" Orientation="Horizontal" Spacing="4" IsVisible="{Binding TimeLeftTillRoundTimeout, Converter={StaticResource TimeSpanVisibilityConverter}}">
-                <TextBlock Text="Registration Ends In:" />
-              </StackPanel>
-              <StackPanel Grid.Row="1" Grid.Column="1" Orientation="Horizontal" Spacing="4" IsVisible="{Binding TimeLeftTillRoundTimeout, Converter={StaticResource TimeSpanVisibilityConverter}}">
-                <TextBlock Foreground="YellowGreen" Text="{Binding TimeLeftTillRoundTimeout, Converter={StaticResource TimeSpanStringConverter}}" />
-              </StackPanel>
-              <StackPanel Grid.Row="2" Grid.Column="0" Orientation="Horizontal" Spacing="4">
-                <TextBlock Text="Number of Registered Peers:" />
-              </StackPanel>
-              <StackPanel Grid.Row="2" Grid.Column="1" Orientation="Horizontal" Spacing="4">
-                <TextBlock Foreground="YellowGreen" Text="{Binding PeersRegistered}" />
-                <TextBlock Text="/" />
-                <TextBlock Text="{Binding PeersNeeded}" />
-                <ProgressBar Minimum="0" Maximum="{Binding PeersNeeded}" Value="{Binding PeersRegistered}" />
-              </StackPanel>
-            </Grid>
-          </StackPanel>
-        </ScrollViewer>
-      </controls:GroupBox>
+      
       <controls:GroupBox Grid.Row="0" Grid.Column="0" Title="CoinJoin" Padding="5" Margin="0 5 5 0">
 
         <Grid Grid.Row="0" Grid.Column="0" Classes="content">
