+++ WalletWasabi.Gui/Controls/WalletExplorer/CoinJoinTabViewModel.cs
@@ -23,14 +23,6 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 	{
 		private CompositeDisposable Disposables { get; set; }
 
-		private long _roundId;
-		private CcjRoundPhase _phase;
-		private DateTimeOffset _roundTimesout;
-		private TimeSpan _timeLeftTillRoundTimeout;
-		private Money _requiredBTC;
-		private string _coordinatorFeePercent;
-		private int _peersRegistered;
-		private int _peersNeeded;
 		private string _password;
 		private Money _amountQueued;
 		private bool _isEnqueueBusy;
@@ -48,7 +40,6 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 			: base("CoinJoin", walletViewModel)
 		{
 			Password = "";
-			TimeLeftTillRoundTimeout = TimeSpan.Zero;
 
 			CoinsList = new CoinListViewModel(Global, CoinListContainerType.CoinJoinTabViewModel);
 
@@ -122,23 +113,6 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 			{
 				CoinJoinUntilAnonimitySet = Global.Config.GetTargetLevel(target);
 			});
-
-			this.WhenAnyValue(x => x.RoundTimesout)
-				.ObserveOn(RxApp.MainThreadScheduler)
-				.Subscribe(x =>
-				{
-					TimeSpan left = x - DateTimeOffset.UtcNow;
-					TimeLeftTillRoundTimeout = left > TimeSpan.Zero ? left : TimeSpan.Zero; // Make sure cannot be less than zero.
-				});
-
-			this.WhenAnyValue(x => x.TimeLeftTillRoundTimeout)
-				.Throttle(TimeSpan.FromSeconds(1))
-				.ObserveOn(RxApp.MainThreadScheduler)
-				.Subscribe(x =>
-				{
-					var next = TimeLeftTillRoundTimeout - TimeSpan.FromSeconds(1);
-					TimeLeftTillRoundTimeout = next > TimeSpan.Zero ? next : TimeSpan.Zero; // Make sure cannot be less than zero.
-				});
 		}
 
 		public override void OnOpen()
@@ -152,37 +126,12 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 
 			TargetPrivacy = Global.Config.GetTargetPrivacy();
 
-			var registrableRound = Global.ChaumianClient.State.GetRegistrableRoundOrDefault();
-
-			UpdateRequiredBtcLabel(registrableRound);
-
-			CoordinatorFeePercent = registrableRound?.State?.CoordinatorFeePercent.ToString() ?? "0.003";
-
 			Observable.FromEventPattern(Global.ChaumianClient, nameof(Global.ChaumianClient.CoinQueued))
 				.Merge(Observable.FromEventPattern(Global.ChaumianClient, nameof(Global.ChaumianClient.CoinDequeued)))
 				.Merge(Observable.FromEventPattern(Global.ChaumianClient, nameof(Global.ChaumianClient.StateUpdated)))
 				.ObserveOn(RxApp.MainThreadScheduler)
 				.Subscribe(_ => UpdateStates())
 				.DisposeWith(Disposables);
-
-			CcjClientRound mostAdvancedRound = Global.ChaumianClient?.State?.GetMostAdvancedRoundOrDefault();
-
-			if (mostAdvancedRound != default)
-			{
-				RoundId = mostAdvancedRound.State.RoundId;
-				Phase = mostAdvancedRound.State.Phase;
-				RoundTimesout = mostAdvancedRound.State.Phase == CcjRoundPhase.InputRegistration ? mostAdvancedRound.State.InputRegistrationTimesout : DateTimeOffset.UtcNow;
-				PeersRegistered = mostAdvancedRound.State.RegisteredPeerCount;
-				PeersNeeded = mostAdvancedRound.State.RequiredPeerCount;
-			}
-			else
-			{
-				RoundId = -1;
-				Phase = CcjRoundPhase.InputRegistration;
-				RoundTimesout = DateTimeOffset.UtcNow;
-				PeersRegistered = 0;
-				PeersNeeded = 100;
-			}
 
 			Global.UiConfig.WhenAnyValue(x => x.LurkingWifeMode).ObserveOn(RxApp.MainThreadScheduler).Subscribe(x =>
 			{
@@ -287,66 +236,6 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 		{
 			AmountQueued = Global.ChaumianClient.State.SumAllQueuedCoinAmounts();
 			MainWindowViewModel.Instance.CanClose = AmountQueued == Money.Zero;
-
-			var registrableRound = Global.ChaumianClient.State.GetRegistrableRoundOrDefault();
-			if (registrableRound != default)
-			{
-				CoordinatorFeePercent = registrableRound.State.CoordinatorFeePercent.ToString();
-				UpdateRequiredBtcLabel(registrableRound);
-			}
-			var mostAdvancedRound = Global.ChaumianClient.State.GetMostAdvancedRoundOrDefault();
-			if (mostAdvancedRound != default)
-			{
-				RoundId = mostAdvancedRound.State.RoundId;
-				if (!Global.ChaumianClient.State.IsInErrorState)
-				{
-					Phase = mostAdvancedRound.State.Phase;
-					RoundTimesout = mostAdvancedRound.State.Phase == CcjRoundPhase.InputRegistration ? mostAdvancedRound.State.InputRegistrationTimesout : DateTimeOffset.UtcNow;
-				}
-				this.RaisePropertyChanged(nameof(Phase));
-				this.RaisePropertyChanged(nameof(RoundTimesout));
-				PeersRegistered = mostAdvancedRound.State.RegisteredPeerCount;
-				PeersNeeded = mostAdvancedRound.State.RequiredPeerCount;
-			}
-		}
-
-#pragma warning disable CS0618 // Type or member is obsolete
-
-		private void UpdateRequiredBtcLabel(CcjClientRound registrableRound)
-#pragma warning restore CS0618 // Type or member is obsolete
-		{
-			if (Global.WalletService is null)
-			{
-				return; // Otherwise NullReferenceException at shutdown.
-			}
-
-			if (registrableRound == default)
-			{
-				if (RequiredBTC == default)
-				{
-					RequiredBTC = Money.Zero;
-				}
-			}
-			else
-			{
-				var queued = Global.WalletService.Coins.Where(x => x.CoinJoinInProgress);
-				if (queued.Any())
-				{
-					RequiredBTC = registrableRound.State.CalculateRequiredAmount(Global.ChaumianClient.State.GetAllQueuedCoinAmounts().ToArray());
-				}
-				else
-				{
-					var available = Global.WalletService.Coins.Where(x => x.Confirmed && !x.Unavailable);
-					if (available.Any())
-					{
-						RequiredBTC = registrableRound.State.CalculateRequiredAmount(available.Where(x => x.AnonymitySet < Global.Config.PrivacyLevelStrong).Select(x => x.Amount).ToArray());
-					}
-					else
-					{
-						RequiredBTC = registrableRound.State.CalculateRequiredAmount();
-					}
-				}
-			}
 		}
 
 		public override void OnSelected()
@@ -389,54 +278,6 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 		{
 			get => _amountQueued;
 			set => this.RaiseAndSetIfChanged(ref _amountQueued, value);
-		}
-
-		public long RoundId
-		{
-			get => _roundId;
-			set => this.RaiseAndSetIfChanged(ref _roundId, value);
-		}
-
-		public CcjRoundPhase Phase
-		{
-			get => _phase;
-			set => this.RaiseAndSetIfChanged(ref _phase, value);
-		}
-
-		public DateTimeOffset RoundTimesout
-		{
-			get => _roundTimesout;
-			set => this.RaiseAndSetIfChanged(ref _roundTimesout, value);
-		}
-
-		public TimeSpan TimeLeftTillRoundTimeout
-		{
-			get => _timeLeftTillRoundTimeout;
-			set => this.RaiseAndSetIfChanged(ref _timeLeftTillRoundTimeout, value);
-		}
-
-		public Money RequiredBTC
-		{
-			get => _requiredBTC;
-			set => this.RaiseAndSetIfChanged(ref _requiredBTC, value);
-		}
-
-		public string CoordinatorFeePercent
-		{
-			get => _coordinatorFeePercent;
-			set => this.RaiseAndSetIfChanged(ref _coordinatorFeePercent, value);
-		}
-
-		public int PeersRegistered
-		{
-			get => _peersRegistered;
-			set => this.RaiseAndSetIfChanged(ref _peersRegistered, value);
-		}
-
-		public int PeersNeeded
-		{
-			get => _peersNeeded;
-			set => this.RaiseAndSetIfChanged(ref _peersNeeded, value);
 		}
 
 		public bool IsEnqueueBusy
